some fixes

+++ src/hnsw.cpp
@@ -1885,9 +1885,9 @@ namespace orangedb {
         int count = 0;
         printf("Number of nodes in last layer: %d\n", storage->fast_level_counters[0]);
         printf("Max neighbors per level: %d\n", storage->max_neighbors_per_level[0]);
-        for (int i = 0; i < storage->fast_level_counters[0]; i++) {
-            for (int j = 0; j < storage->max_neighbors_per_level[0]; j++) {
-                if (nbrs[i * storage->max_neighbors_per_level[0] + j] != INVALID_VECTOR_ID) {
+        for (size_t i = 0; i < storage->fast_level_counters[0]; i++) {
+            for (size_t j = 0; j < storage->max_neighbors_per_level[0]; j++) {
+                if (nbrs[i * (size_t) storage->max_neighbors_per_level[0] + j] != INVALID_VECTOR_ID) {
                     count++;
                 }
             }
